Fixing settings not saving
Multiple changes to be inline with other plugins.
Correcting settings file name to allow updating.
Dereferencing options variable class to allow for future stability.
