+++ Jellyfin.Plugin.PlaybackReporting/Model/ReportPlaybackOptions.cs
@@ -14,12 +14,12 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see<http://www.gnu.org/licenses/>.
 */
 
-using MediaBrowser.Model.Plugins;
-
 namespace Jellyfin.Plugin.PlaybackReporting
 {
-    public class PluginConfiguration: BasePluginConfiguration
+    public class ReportPlaybackOptions
     {
-
+        public int MaxDataAge { set; get; } = 3;
+        public string BackupPath { set; get; }
+        public int MaxBackupFiles { set; get; } = 5;
     }
 }
