+++ Jellyfin.Plugin.PlaybackReporting/Plugin.cs
@@ -20,6 +20,7 @@ using MediaBrowser.Common.Configuration;
 using MediaBrowser.Common.Plugins;
 using MediaBrowser.Model.Plugins;
 using MediaBrowser.Model.Serialization;
+using Jellyfin.Plugin.PlaybackReporting.Configuration;
 
 namespace Jellyfin.Plugin.PlaybackReporting
 {
